+++ HISTORY.md
@@ -25,6 +25,8 @@ http://visjs.org
 - Fixed newly added item ignored when returning an other object instance.
 - Fixed option `autoResize` not working on IE in case of changing visibility
   of the Timeline container element.
+- Renamed internal items from `ItemBox`, `ItemRange`, and `ItemPoint` to
+  respectively `BoxItem`, `RangeItem`, and `PointItem`.
 
 
 ## 2014-08-29, version 3.3.0